requirements.txt
# lab1
spade

# misc
ipykernel
python-dotenv
